python really helps clean things up
